Add files via upload
This is a case study that captures the exploratory data analysis of the largest US American airlines. Key questions answered are (1) What are the most important business insights you find to understand the current state of the industry? (2) What are any interesting trends, performance drivers, events etc. in the airline sector in the past 19 years?